Reorganize common cluster unit tests
